+++ CHANGELOG.md
@@ -106,12 +106,12 @@
 
 ### Bug Fixes
 
-* update import handling and type casting in JWT sign function ([#6](https://github.com/AntelopeJS/auth/issues/6)) ([a89b5a4](https://github.com/AntelopeJS/auth/commit/a89b5a46d854761e5f5bb10cd864d27c796c2d43))
+- update import handling and type casting in JWT sign function ([#6](https://github.com/AntelopeJS/auth/issues/6)) ([a89b5a4](https://github.com/AntelopeJS/auth/commit/a89b5a46d854761e5f5bb10cd864d27c796c2d43))
 
 ## [0.1.0](https://github.com/AntelopeJS/auth/compare/v0.0.1...v0.1.0) (2025-05-29)
 
 ### Features
 
-* default config ([#4](https://github.com/AntelopeJS/auth/issues/4)) ([ace5215](https://github.com/AntelopeJS/auth/commit/ace52152ba9d4370a82e612161c6e2ed310161a2))
+- default config ([#4](https://github.com/AntelopeJS/auth/issues/4)) ([ace5215](https://github.com/AntelopeJS/auth/commit/ace52152ba9d4370a82e612161c6e2ed310161a2))
 
 ## 0.0.1 (2025-05-08)